+++ include/slepcbv.h
@@ -184,6 +184,7 @@ SLEPC_EXTERN PetscErrorCode BVNormColumnEnd(BV,PetscInt,NormType,PetscReal*);
 SLEPC_EXTERN PetscErrorCode BVNormalize(BV,PetscScalar*);
 SLEPC_EXTERN PetscErrorCode BVSetRandom(BV);
 SLEPC_EXTERN PetscErrorCode BVSetRandomNormal(BV);
+SLEPC_EXTERN PetscErrorCode BVSetRandomSign(BV);
 SLEPC_EXTERN PetscErrorCode BVSetRandomColumn(BV,PetscInt);
 SLEPC_EXTERN PetscErrorCode BVSetRandomCond(BV,PetscReal);
 SLEPC_EXTERN PetscErrorCode BVSetRandomContext(BV,PetscRandom);